refactor Feeds page to make smaller
